Move truncate funciton to utils/operations
