+++ src/UI/TransactionInfo/components/DataField/DataField.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { N_A } from 'constants/index';
 import { ScamInfoType } from 'types/account.types';
 import { DataDecode } from 'UI/TransactionInfo/components/DataDecode/index';
+import { truncate } from 'utils/operations/truncate';
 import { getScamFlag } from 'utils/transactions/transactionInfoHelpers/getScamFlag';
 import { useDataDecodeMethod } from 'utils/transactions/transactionInfoHelpers/useDataDecodeMethod';
 import { Anchorme, ModalLink } from './components';
-import { truncate } from './helpers/truncate';
 
 const DISPLAYED_DATA_LENGTH = 1000000;
