--- src/UI/TransactionInfo/components/DataField/helpers/index.tsx
@@ -1,1 +1,0 @@
-export * from './truncate';
