+++ src/utils/operations/truncate.tsx
